build: add App component
Although I previously removed this I was thinking at the time where I'd store specific state that needs to persist with the navbar, and thus App has been created again.

build: move StyledNavbar and RouteProvider here

By moving these components into this component and calling this component in main, we can store state in this component and pass it to the StyledNavbar later so that component can display a count of items in the cart in real time.

style: apply a basic style via styled