feat: add App component
Previously this project had a starting App component, which I removed because I didn't think it was needed until I started thinking about state variables. App will be the main component and items pulled from the api and items added to the users shopping cart will exist within this component and be passed down to shopping page and cart page components.

feat: add StyledNavbar and RouteProvider

StyledNavbar will be persistent no matter which of the three pages of the app the user navigates to, so rendering it in App is needed. RouteProvider will then replace the main content of the page with either Homepage, Shoppage, or Cartpage depending on route navigated to.

style: add styled usage to style

This is a temporary style rule just to practice styled, I don't think I'll actually style App in the future and it will just be used to set up state variables and fetch store items, but we'll see.