test: change test case descriptive text
I've changed the text content that gets rendered in a NavBar component and just want this descriptive text to reflect that change.

test: add screen.debug call

This allows us to see the rendered html when App renders NavBar

test: add test unit to look for "Nav" in the doc

This test unit is created to specifically test whether the content in the NavBar is rendered in the document, which should be rendered in every page as this is a persistent component anyways.