feat: add useEffect to pull in api data
This api fetch on mount only obtains the product objects from the fakestore api.

feat: add prop to NavBar component rendering

This prop of cartItems.length will allow us to pass into the NavBar the number of items in the users shopping cart and with that we'll display in the future on every page how many items the user has in their cart.