feat: add Link elements wrapping new svg's
These svg's will become clickable navigation icons for the users to navigate to the other pages of the application.

style: add flex to the component, style svg's

By adding flex and declaring spacing we have a dynamic  layout for our NavBar and I've set a limit on these icons width and height's so they don't get too big.