build: import useEffect, useState, and Outlet
NavBar is now the primary component of this application, therefore the state will reside in it and the fetch to get products on mount will also reside in it using useEffect.

feat: add headings to navigation icons

This should help users who don't perceive the meaning of the icons to instead be able to read the headings.

style: add styled-components styling

Styling has been added to the many elements to appropriately position dynamically.