feat: rewrite root to StyledNavBar
NavBar is now the root component, and such...

feat: add children routes to the root route

In addition to setting up a default child component of StyledHomePage, I've set up the appropriate roots for shop and cart pages. This way the NavBar is persistent and passes the state down to the other components.