build: import useOutletContext
This component is now a child of NavBar and for its parent to pass down state we'll need to use context attribute in Outlet in the parent and  import it into the component by destructuring with useOutletContext.

feat: destructure state from context of Outlet

This state variable and its state setting function will allow us to display items in a users shopping card, and to edit them in quantity via the state's setter function.