build: import useOutletContext
This will be used to pull in state from NavBar.

feat: add data to component for products

This entails a header with the product title, and image of the product provided by the api, the rating of the product by users from the api, a product description, product price, and finally an input field and button to add a quantity of the product to the shopping cart of the user.

style: add styling to elements