feat: add pros for cartItems and setCartItems
cartItems passes into ProductCard the state of the shopping cart persay and passing in its setting function allows us to add onClick ability to the add to cart button to add the user entered quantity of a product to their cart.

feat: add span of 'Quantity'

This is placed before the input element so users will know exactly what that input element is for.

feat: add onClick attribute and logic to button

The logic of this button when clicked either adds an object to the cartItems array or just increases the appropriate cart item's quantity.