feat: add CartCard component
This component will be used in CartPage to communicate each product a user has added to their shopping cart.

build: import styled-components styled method

This will allow us to style the component in its own js file.

feat: add props

In addition to the usual project state stored in NavBar we're also importing cartItem, which specifically communicates each product in cartItems via id and quantity property value pairs.

feat: add className and key to returned div

These allow styled-components to apply styling to the div element that is returned by this component and tie the element to an object in state so that React will know when it needs to rerender this element when that state is changed.

feat: add info elements and interpolate data

I've added a heading, an image, input elements, and paragraph elements to communicate information about each product a user has added to their cart and how much each unit is and how much the combined unit cost is for that product.

style: add styling

I've styled the returned element to match my planned layout of data of the component which hopefully would be easy for the user to understand the product in their cart.

feat: add delete button

The logic is missing, but this button will allow users to remove a product from their shopping carts.