feat: add totalCartPrice variable
This variable's value will be interpolated below all CartCard/product elements to display the total cost of the shopping cart contents to the user.

feat: add checkout button

This is just for looks and does not actually do anything.

feat: add dynamic content

I added an if else conditional statement that checks the number of items in the cartItems state and if there are any items then returns a header and a CartCard component for each product in cartItems and uses the totalCartPrice to communicate their shopping cart total price, otherwise returns a paragraph element that tells the user their cart is empty.