style: change the NavBar to be sticky
This allows the NavBar component to stay at the top of the users browser on pages like '/shop' where there are 20 products displayed and scrolling down the page is very apparent. I wanted the cart item count to be displayed at all times and this accomplishes that.

refactor: remove commented out unnecessary code

I had a state setting function call in the useEffect that I was pondering but found a better way to set up that state variables value.