feat: add data-testid's to returned elements
These will be used in testing to verify that these components are being rendered.

fix: move up the state to App

NavBar was incorrectly storing the state and I've now moved it up to App which is the new root component which creates an instance of NavBar and then below it the appropriate component depending on user navigation.

fix: move up fetch call to App

As state has been moved to App and App is now the root component the fetch/useEffect on mount needs to reside in App.