style: move dynamic className to root element
Before this change the className applied by styled-components was being applied to the nav, which was not the root element returned by this component and caused the NavBar returned elements to not be sticky. Also applied styling to nav element now that it is not the initial targeted element by styled-components so match prior planned design.