style: center align the text in the paragraph
This paragraph element holds the item descriptions and was ugly prior to this change.

style: change multiple spans to be bolded

This now affects the rating content within a span to be bolded as well as the original bolded price span.

style: apply a specific height to product images

Before this chang ProductCards would load and then after a half second change in size to paint the image into the component. This change sets aside height for the image at component mounting so no more readjusting happens.