fix: change span to label
This label now wraps the input element of ProductCard instances which allows the user to enter the input element when clicking the label, and gives semantic value this element as well.

fix: change onClick functionality

The changes to this action now show an error message when a user has not entered a quantity of an item to add to their carts but have clicked on the 'Add to Cart' button. Before these changes there was no message and there was code executing to change nothing in the background, which won't invoke now until there is an input value greater than zero. This error message also is not spammable which was prevented by adding a variable to the component scope and using a condition off that value before showing that error message.