feat: add error message for when user enters <1
This pertains to when a user updates the quantity of an item in their cart. We'd like them to not checkout with a bunch of zero quantity items in their cart, so we display them a message telling them that if they want to remove a product from their cart to instead press the delete button.