refactor: move assets out of public into root
refactor: rename assets to public

public is the generally used name for the directory that holds assets other than code.

refactor: move router file to src

This file was previously in assets which is now public but that directory is not for code.

refactor: change imports to reflect moved files